NODE-99, refactor: use multi-sig primitive
