+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
 	"precompiles/btc-registration-pool",
 	"primitives/core",
 	"primitives/bfc-staking",
+	"primitives/multi-sig",
 ]
 [profile.release]
 panic = "unwind"
@@ -79,6 +80,7 @@ array-bytes = "6.1"
 # Bifrost Primitive
 bp-core = { default-features = false, path = "primitives/core" }
 bp-staking = { default-features = false, path = "primitives/bfc-staking" }
+bp-multi-sig = { default-features = false, path = "primitives/multi-sig" }
 
 # Bifrost Runtime
 bifrost-common-constants = { default-features = false, path = "runtime/common/constants" }
