+++ precompiles/btc-registration-pool/Cargo.toml
@@ -12,6 +12,7 @@ repository = { workspace = true }
 # Bifrost
 pallet-btc-registration-pool = { workspace = true }
 precompile-utils = { workspace = true }
+bp-multi-sig = { workspace = true }
 
 # Substrate
 parity-scale-codec = { workspace = true, features = ["max-encoded-len"] }
@@ -40,4 +41,5 @@ std = [
 	"sp-core/std",
 	"sp-std/std",
 	"sp-runtime/std",
+	"bp-multi-sig/std",
 ]
