+++ primitives/multi-sig/src/lib.rs
@@ -1,0 +1,106 @@
+pub mod traits;
+
+use sp_core::{ConstU32, RuntimeDebug};
+use sp_runtime::{BoundedBTreeMap, BoundedVec};
+use sp_std::vec::Vec;
+
+use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
+use scale_info::TypeInfo;
+
+/// The maximum amount of accounts a multi-sig account can consist.
+pub const MULTI_SIG_MAX_ACCOUNTS: u32 = 16;
+
+/// The maximum length of a valid Bitcoin address in characters (~62 alphanumeric characters).
+pub const ADDRESS_MAX_LENGTH: u32 = 62;
+
+/// The Bitcoin address type (length bounded).
+pub type BoundedBitcoinAddress = BoundedVec<u8, ConstU32<ADDRESS_MAX_LENGTH>>;
+
+#[derive(
+	Clone,
+	Copy,
+	Decode,
+	Encode,
+	Eq,
+	Ord,
+	PartialOrd,
+	PartialEq,
+	TypeInfo,
+	MaxEncodedLen,
+	RuntimeDebug,
+)]
+/// A 33 byte length public key.
+pub struct Public(pub [u8; 33]);
+
+impl AsRef<[u8]> for Public {
+	fn as_ref(&self) -> &[u8] {
+		&self.0
+	}
+}
+
+#[derive(Decode, Encode, TypeInfo, MaxEncodedLen)]
+/// A m-of-n multi signature based Bitcoin address.
+pub struct MultiSigAccount<AccountId> {
+	/// The vault address.
+	pub address: AddressState,
+	/// Public keys that the vault address contains.
+	pub pub_keys: BoundedBTreeMap<AccountId, Public, ConstU32<MULTI_SIG_MAX_ACCOUNTS>>,
+	/// The m value of the multi-sig address.
+	pub m: u8,
+	/// The n value of the multi-sig address.
+	pub n: u8,
+}
+
+impl<AccountId: PartialEq + Clone + Ord> MultiSigAccount<AccountId> {
+	pub fn new(m: u8, n: u8) -> Self {
+		Self { address: AddressState::Pending, pub_keys: BoundedBTreeMap::new(), m, n }
+	}
+
+	pub fn is_pending(&self) -> bool {
+		matches!(self.address, AddressState::Pending)
+	}
+
+	pub fn is_address(&self, address: &BoundedBitcoinAddress) -> bool {
+		match &self.address {
+			AddressState::Pending => false,
+			AddressState::Generated(a) => a == address,
+		}
+	}
+
+	pub fn is_key_generation_ready(&self) -> bool {
+		self.n as usize == self.pub_keys.len()
+	}
+
+	pub fn is_key_submitted(&self, pub_key: &Public) -> bool {
+		self.pub_keys.values().cloned().collect::<Vec<Public>>().contains(pub_key)
+	}
+
+	pub fn is_authority_submitted(&self, authority_id: &AccountId) -> bool {
+		self.pub_keys.contains_key(authority_id)
+	}
+
+	pub fn insert_pub_key(
+		&mut self,
+		authority_id: AccountId,
+		pub_key: Public,
+	) -> Result<Option<Public>, (AccountId, Public)> {
+		self.pub_keys.try_insert(authority_id, pub_key)
+	}
+
+	pub fn set_address(&mut self, address: BoundedBitcoinAddress) {
+		self.address = AddressState::Generated(address)
+	}
+
+	pub fn pub_keys(&self) -> Vec<Public> {
+		self.pub_keys.values().cloned().collect()
+	}
+}
+
+#[derive(Decode, Encode, TypeInfo, MaxEncodedLen)]
+/// The vault address state.
+pub enum AddressState {
+	/// Required number of public keys has not been submitted yet.
+	Pending,
+	/// n public keys has been submitted and address generation done.
+	Generated(BoundedBitcoinAddress),
+}
